fix: cancel Android playback echo in JNI pipeline (#11562)
