+++ packages/shared/src/voice/aec/index.ts
@@ -1,0 +1,18 @@
+export {
+  DEFAULT_PLAYBACK_DELAY_MS,
+  type EchoDelayEstimate,
+  type EchoDelayOptions,
+  estimateEchoDelaySamples,
+  PLATFORM_PLAYBACK_DELAY_DEFAULTS,
+  platformPlaybackDelayMs,
+  platformPlaybackDelaySamples,
+} from "./echo-delay.js";
+export {
+  EchoReferenceBuffer,
+  type EchoReferenceBufferOptions,
+} from "./echo-reference-buffer.js";
+export {
+  NlmsEchoCanceller,
+  type NlmsEchoCancellerOptions,
+  type ResidualSuppressionOptions,
+} from "./nlms-echo-canceller.js";
